requirements.txt
Django==3.2.3
djangorestframework==3.12.2
model-bakery==1.2.1
django-cors-headers==3.5.0
Pillow==8.4.0
